refactor: app.js 함수 별도 분리
