+++ week12/sebellko/modules/captureByTimeOut.js
@@ -1,0 +1,13 @@
+import setTimer from './setTimer.js';
+
+const captureCanvasByTimeOut = async (target, time) => {
+  const { timeOut } = await setTimer(minute, second, time);
+  if (timeOut) {
+    canvas.toBlob((blob) => {
+      const url = URL.createObjectURL(blob);
+      target.src = url;
+    });
+  }
+};
+
+export default captureCanvasByTimeOut;
