+++ week12/sebellko/style.css
@@ -29,6 +29,7 @@ main {
 .player_list {
   width: 100%;
   display: flex;
+  align-items: center;
   justify-content: center;
   flex-wrap: wrap;
   gap: 50px;
